+++ apalis-core/src/backend/custom.rs
@@ -29,10 +29,11 @@
 //! # use futures_util::StreamExt;
 //! # use futures_util::FutureExt;
 //! # use apalis_core::backend::TaskSink;
+//! # use apalis_core::task::task_id::RandomId;
 //! #[tokio::main]
 //! async fn main() {
 //!     // Create a memory-backed VecDeque
-//!     let memory = Arc::new(Mutex::new(VecDeque::<Task<u32, ()>>::new()));
+//!     let memory = Arc::new(Mutex::new(VecDeque::<Task<u32, (), RandomId>>::new()));
 //!
 //!     // Build the custom backend
 //!     let mut backend = BackendBuilder::new()
@@ -403,6 +404,7 @@ mod tests {
 
     use crate::{
         error::BoxDynError,
+        task::task_id::RandomId,
         worker::{builder::WorkerBuilder, ext::event_listener::EventListenerExt},
     };
 
@@ -412,7 +414,8 @@ mod tests {
 
     #[tokio::test]
     async fn basic_custom_backend() {
-        let memory: Arc<Mutex<VecDeque<Task<u32, ()>>>> = Arc::new(Mutex::new(VecDeque::new()));
+        let memory: Arc<Mutex<VecDeque<Task<u32, (), RandomId>>>> =
+            Arc::new(Mutex::new(VecDeque::new()));
 
         let mut backend = BackendBuilder::new()
             .database(memory)